# app/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FEM 巡檢管理系統 API",
    "description": "設施設備巡檢管理系統的 RESTful API 文件",
    "version": "1.0.0",
    "contact": {
      "name": "API 支援",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API 伺服器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "輸入您的 JWT Token"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "full_name": {
            "type": "string",
            "example": "系統管理員"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "role": {
            "type": "object",
            "properties": {
              "role_id": {
                "type": "integer"
              },
              "role_name": {
                "type": "string"
              }
            }
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InspectionTask": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "integer"
          },
          "task_name": {
            "type": "string"
          },
          "task_date": {
            "type": "string",
            "format": "date"
          },
          "shift": {
            "type": "string",
            "enum": ["早班", "中班", "夜班"]
          },
          "route": {
            "type": "object",
            "properties": {
              "route_id": {
                "type": "string"
              },
              "route_name": {
                "type": "string"
              }
            }
          },
          "assigned_to": {
            "$ref": "#/components/schemas/User"
          },
          "status": {
            "type": "string",
            "enum": ["未開始", "進行中", "已完成", "已逾期"]
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/tasks/download": {
      "get": {
        "tags": ["任務管理"],
        "summary": "下載巡檢任務",
        "description": "下載指派給使用者的巡檢任務（含完整路線資料）",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "使用者 ID（可選，預設為當前使用者）",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "指定日期（格式：YYYY-MM-DD）",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功下載任務",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tasks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/InspectionTask"
                          }
                        },
                        "last_sync": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/list": {
      "get": {
        "tags": ["任務管理"],
        "summary": "取得任務列表",
        "description": "查詢巡檢任務列表",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["未開始", "進行中", "已完成", "已逾期"]
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得任務列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tasks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/InspectionTask"
                          }
                        },
                        "pagination": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}